backend/requirements.txt
flask==3.0.3
flask-cors==3.1.1
psycopg[binary]==3.2.1
bcrypt==4.0.1
PyJWT==2.8.0
gunicorn==22.0.0
pytest==8.3.2